create,read,delete is working.

+++ expense-tracker-app/controller/groups/destroy.php
@@ -2,12 +2,18 @@
 
 
 use Core\Database\App;
-$id=$_POST['id'];
+$id = $_POST['groupId'];
+
+if (!$id) {
+    echo json_encode(["success" => false, "error" => "Invalid group ID"]);
+    exit();
+}
+
 $db=App::resolve('Core\Database\Database');
 
 $db->query('delete from `groups` where id=:id',[
       'id'=>$id,
 ]);
 
-header("Location: /");
+echo json_encode(["success" => true]);
 exit();
